requirements.txt
networkx
torch
torchvision
torch-summary
einops
scikit-image
opencv-python
pyclipper
shapely
requests
git+https://github.com/wbond/oscrypto
cryptography
freetype-py
googletrans==4.0.0rc1
aiohttp
tqdm
deepl
ImageHash
kornia
backports.cached-property
huggingface_hub
transformers
py3langid
sentencepiece
editdistance
numpy
tensorboardX
websockets
protobuf
ctranslate2
colorama
openai==0.28
open_clip_torch
safetensors
pandas
onnxruntime
omegaconf
python-dotenv
nest-asyncio
marshmallow
cython
aioshutil
aiofiles
arabic-reshaper
pyhyphen
langcodes
llama-cpp-python